requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==3.1.12
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djongo==1.3.6
dlib==19.24.6
dnspython==2.7.0
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.10
numpy==2.2.1
opencv-python==4.11.0.86
pillow==11.1.0
psycopg2==2.9.10
PyJWT==1.7.1
pymongo==3.11.4
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.2.4
tzdata==2024.2
urllib3==2.3.0
